All the internal iterators moved to separate namespace.
